+++ Spaghettis/Modules/spaghettis_gui/GUI/Patches/Graphics/Painters/BoxPainter.hpp
@@ -28,10 +28,10 @@ public:
 // MARK: -
 
 public:
-    void paintObject (juce::Rectangle<float>, juce::Graphics&) override;
+    void paintObject (juce::Rectangle<int>, juce::Graphics&) override;
 
 public:
-    juce::Rectangle<float> getRequiredBoundsForObject() override;
+    juce::Rectangle<int> getRequiredBoundsForObject() override;
 
 private:
     juce::String getText() const;
